cleanups and adjust learning rate for fp16

+++ extra/models/retinanet.py
@@ -143,14 +143,14 @@ class ClassificationHead:
 
     if Tensor.training:
       assert labels is not None and matches is not None, "labels and matches should be passed in when training"
-      return self._compute_loss(out, labels, matches)
+      return self._compute_loss(out.cast(dtypes.float32), labels, matches)
 
     return out.sigmoid()
 
   def _compute_loss(self, x:Tensor, labels:Tensor, matches:Tensor) -> Tensor:
     labels = ((labels + 1) * (fg_idxs := matches >= 0) - 1).one_hot(num_classes=x.shape[-1])
     valid_idxs = (matches != -2).reshape(matches.shape[0], -1, 1)
-    loss = valid_idxs.where(sigmoid_focal_loss(x.cast(dtypes.float32), labels), 0).sum(-1).sum(-1)
+    loss = valid_idxs.where(sigmoid_focal_loss(x, labels), 0).sum(-1).sum(-1)
     loss = (loss / fg_idxs.sum(-1)).sum() / matches.shape[0]
     return loss
 
@@ -175,7 +175,7 @@ class RegressionHead:
 
   def _compute_loss(self, x:Tensor, bboxes:Tensor, matches:Tensor, anchors:Tensor) -> Tensor:
     mask = (fg_idxs := matches >= 0).reshape(matches.shape[0], -1, 1)
-    x = x.cast(dtypes.float32) * mask
+    x = x * mask
     tgt = self.box_coder.encode(bboxes, anchors) * mask
     loss = l1_loss(x, tgt).sum(-1).sum(-1)
     loss = (loss / fg_idxs.sum(-1)).sum() / matches.shape[0]
